# openapi.yaml
openapi: 3.0.3
info:
  title: Laundry Pickup & Delivery API
  version: 1.3.0
  license:
    name: Academic project
    url: https://opensource.org/licenses/MIT
  description: >-
    Contract-first API for submitting laundry orders and arranging pickup.
    Clients MUST ignore unknown response fields and MUST safely handle unknown
    enumeration values without crashing. OAuth 2.0 / OIDC authentication and
    scope-based authorization are required for all protected operations.
servers:
  - url: https://pbse.kevinio.my.id/v1
    description: Deployed service
  - url: http://localhost:8080/v1
    description: Local service
security:
  - oauth2: []
...
paths:
...
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get one laundry order
      operationId: getOrder
      security:
        - oauth2: [orders:read]
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Laundry order
          headers:
            ETag:
              description: >-
                Opaque version marker for this order. Send it back as
                `If-None-Match` to poll cheaply, and as `If-Match` on a write
                that must not overwrite somebody else's change. Strong
                validator: it never carries the `W/` prefix, because
                `If-Match` requires strong comparison.
              schema:
                type: string
                example: '"3f9c1a7b8e2d4f60a1b2c3d4e5f60718"'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              example:
                id: ord_01HZX2Y5K7
                customerId: cus_01HZX2Y1AB
                serviceType: wash_fold
                weightKg: 5.5
                pickupAddress: Jl. Merdeka No. 10, Jakarta
                status: pending_pickup
                createdAt: '2026-09-02T10:00:00Z'
                updatedAt: '2026-09-02T10:00:00Z'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
...
  /pickups:
    get:
      tags: [Pickups]
      summary: List pickup arrangements
      operationId: listPickups
      description: >-
        Returns the pickups the caller is allowed to see. A driver sees only
        their own assignments; staff see the pickups of orders belonging to
        their outlet.
      security:
        - oauth2: [pickups:read]
      parameters:
        - $ref: '#/components/parameters/PickupStatusFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Pickup collection
          headers:
            ETag:
              description: >-
                Version marker for this page. Send it back as `If-None-Match`
                on the next poll to receive `304` when nothing changed.
              schema:
                type: string
            X-Next-Cursor:
              description: >-
                Opaque cursor for the next page. Present only when more rows
                exist. Pass it back in the `cursor` query parameter.
              schema:
                type: string
                example: eyJjIjoiMjAyNi0wOS0wM1QwOTowMDowMC4wMDBaIiwiaSI6InBrdV8wMUhaWDJaOEs5In0
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pickup'
              example:
                - id: pku_01HZX2Z8K9
                  orderId: ord_01HZX2Y5K7
                  driverId: drv_01HZX2Z1AA
                  scheduledAt: '2026-09-03T09:00:00Z'
                  address: Jl. Merdeka No. 10, Jakarta
                  status: scheduled
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
    post:
      tags: [Pickups]
      summary: Dispatch a pickup for an order
      operationId: createPickup
      description: >-
        Staff-only operation used to assign a driver to an order's pickup. The
        service loads the order named by `orderId` and answers with the same
        `404` used for an unknown order when that order is not within the
        caller's outlet, so identifiers cannot be enumerated. Idempotency-Key
        follows the same 24-hour contract as `createOrder`. The created pickup
        starts in `assigned` status because a driver is named at creation time.
      security:
        - oauth2: [orders:fulfil]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePickupRequest'
            example:
              orderId: ord_01HZX2Y5K7
              driverId: drv_01HZX2Z1AA
              scheduledAt: '2026-09-03T09:00:00Z'
      responses:
        '201':
          description: Pickup dispatched
          headers:
            Location:
              description: URL of the newly created pickup.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pickup'
              example:
                id: pku_01HZX2Z8K9
                orderId: ord_01HZX2Y5K7
                driverId: drv_01HZX2Z1AA
                scheduledAt: '2026-09-03T09:00:00Z'
                address: Jl. Merdeka No. 10, Jakarta
                status: assigned
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /pickups/{pickupId}/collect:
    post:
      tags: [Pickups]
      summary: Mark an assigned pickup as collected
      operationId: collectPickup
      description: >-
        Driver operation. Only the driver named in the pickup's `driverId` may
        collect it; any other caller receives the same `404` as an unknown
        pickup. The ownership check runs before the status is written, so a
        refused request never mutates the record. The transition is naturally
        idempotent: collecting an already collected pickup returns the current
        representation with `200` and does not require an Idempotency-Key.
      security:
        - oauth2: [pickups:write]
      parameters:
        - $ref: '#/components/parameters/PickupId'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Pickup collected, or already collected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pickup'
              example:
                id: pku_01HZX2Z8K9
                orderId: ord_01HZX2Y5K7
                driverId: drv_01HZX2Z1AA
                scheduledAt: '2026-09-03T09:00:00Z'
                address: Jl. Merdeka No. 10, Jakarta
                status: picked_up
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: OIDC / OAuth 2.0 Authorization Server authentication and authorization
      flows:
        authorizationCode:
          authorizationUrl: https://auth.example.com/authorize
          tokenUrl: https://auth.example.com/oauth/token
          scopes:
            'orders:read': Read order collection and details within caller authority
            'orders:write': Create and cancel orders owned by caller
            'pickups:read': Read pickup collection and details within caller authority
            'orders:fulfil': Manage order processing and fulfilment (staff)
            'pickups:write': Update assigned pickup status (driver)
        clientCredentials:
          tokenUrl: https://auth.example.com/oauth/token
          scopes:
            'orders:read': Read order collection and details for system jobs
            'pickups:read': Read pickup collection and details for system jobs
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      description: Opaque server-generated order identifier
      schema:
        type: string
        pattern: '^ord_[A-Za-z0-9]+$'
        example: ord_01HZX2Y5K7
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of resources returned in the collection.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor for retrieving the next collection page.
      schema:
        type: string
        example: cur_01HZX31ABC
    PickupId:
      name: pickupId
      in: path
      required: true
      description: Opaque server-generated pickup identifier
      schema:
        type: string
        pattern: '^pku_[A-Za-z0-9]+$'
        example: pku_01HZX2Z8K9
    PickupStatusFilter:
      name: status
      in: query
      required: false
      description: Return pickups with this status
      schema:
        type: string
        enum: [scheduled, assigned, picked_up, delivered, cancelled]
        example: scheduled
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Canonical UUID v4 generated when the user confirms the operation and
        persisted with the pending request. The same key must be reused for
        every retry of that request and is retained for 24 hours. A new key
        after retention is treated as a new request. The same key with the
        same request returns the original result; the same key with different
        request data returns 409 Conflict.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
        example: 423e4567-e89b-42d3-a456-426614174000
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: >-
        Version marker of the entity as the client last saw it, taken from the
        `ETag` of a preceding read. When it is present and does not match the
        current version, the service answers `412` instead of applying the
        write, so a client cannot overwrite a change it has not seen.

        Optional, and deliberately so: a client that omits it keeps the
        pre-1.3.0 behaviour. A client that wants the lost-update guarantee must
        send it. Strong comparison applies, so a weak tag never matches.
      schema:
        type: string
        example: '"3f9c1a7b8e2d4f60a1b2c3d4e5f60718"'
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: >-
        Version marker from a previous read of the same resource. When it
        matches the current version the service answers `304` with no body,
        which is what makes a poll cheap.
      schema:
        type: string
        example: '"3f9c1a7b8e2d4f60a1b2c3d4e5f60718"'
  schemas:
    Order:
      type: object
      required: [id, customerId, serviceType, weightKg, pickupAddress, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          pattern: '^ord_[A-Za-z0-9]+$'
          example: ord_01HZX2Y5K7
        customerId:
          type: string
          pattern: '^cus_[A-Za-z0-9]+$'
          example: cus_01HZX2Y1AB
        serviceType:
          type: string
          enum: [wash_fold, dry_clean, iron_only]
          example: wash_fold
        weightKg:
          type: number
          format: float
          minimum: 0.1
          example: 5.5
        pickupAddress:
          type: string
          example: Jl. Merdeka No. 10, Jakarta
        status:
          type: string
          enum: [pending_pickup, ready_for_pickup, confirmed, assigned, picked_up, processing, completed, cancelled]
          description: Clients MUST safely handle unknown values as an in-progress state.
          example: pending_pickup
        outletId:
          type: string
          nullable: true
          description: >-
            Outlet that took this order in. Null until a staff member accepts
            the order; added in v1.1.0 and safe for clients to ignore.
          example: outlet_a
        createdAt:
          type: string
          format: date-time
          example: '2026-09-02T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-09-02T10:00:00Z'
...
    Pickup:
      type: object
      required: [id, orderId, driverId, scheduledAt, address, status]
      properties:
        id:
          type: string
          pattern: '^pku_[A-Za-z0-9]+$'
          example: pku_01HZX2Z8K9
        orderId:
          type: string
          pattern: '^ord_[A-Za-z0-9]+$'
          example: ord_01HZX2Y5K7
        driverId:
          type: string
          pattern: '^drv_[A-Za-z0-9]+$'
          example: drv_01HZX2Z1AA
        scheduledAt:
          type: string
          format: date-time
          example: '2026-09-03T09:00:00Z'
        address:
          type: string
          example: Jl. Merdeka No. 10, Jakarta
        status:
          type: string
          enum: [scheduled, assigned, picked_up, delivered, cancelled]
          example: scheduled
    CreatePickupRequest:
      type: object
      required: [orderId, driverId, scheduledAt]
      properties:
        orderId:
          type: string
          pattern: '^ord_[A-Za-z0-9]+$'
          example: ord_01HZX2Y5K7
        driverId:
          type: string
          pattern: '^drv_[A-Za-z0-9]+$'
          example: drv_01HZX2Z1AA
        scheduledAt:
          type: string
          format: date-time
          example: '2026-09-03T09:00:00Z'
...
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri-reference
          example: https://api.example.com/problems/order-not-cancellable
        title:
          type: string
          example: Order cannot be cancelled
        status:
          type: integer
          format: int32
          example: 409
        detail:
          type: string
          example: Order has already entered processing.
        instance:
          type: string
          example: /v1/orders/ord_01HZX2Y5K7/cancellation
        currentStatus:
          type: string
          example: assigned
        allowedStatuses:
          type: array
          items:
            type: string
          example: [pending_pickup, ready_for_pickup, confirmed]
        invalidFields:
          type: array
          description: Names of request fields that failed validation.
          items:
            type: string
          example: [serviceType, weightKg]
        invalid-params:
          type: array
          description: >-
            Machine-readable list of the request fields that failed validation,
            each with the reason it failed. Extension member in the shape
            RFC 9457 uses in its own example. Sent alongside `invalidFields`,
            which is retained for clients written against earlier versions.
          items:
            type: object
            required: [name, reason]
            properties:
              name:
                type: string
                example: weightKg
              reason:
                type: string
                example: must be at least 0.1
        error:
          type: string
          description: >-
            OAuth 2.0 error code, present on `401` responses and mirrored in
            the `WWW-Authenticate` header.
          example: invalid_token
        requiredScopes:
          type: array
          description: Scopes the operation required, present on `403` responses.
          items:
            type: string
          example: [orders:fulfil]
  responses:
    NotModified:
      description: >-
        The resource is unchanged since the version named in `If-None-Match`.
        The body is empty by definition; a client treats this as a successful
        read and clears any stale marker.
    PreconditionFailed:
      description: >-
        The entity changed since the version named in `If-Match`. A normal
        condition, not a system failure: the client is expected to reload,
        re-render, and explain it in domain terms rather than show a generic
        error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/precondition-failed
            title: Precondition failed
            status: 412
            detail: >-
              Order ini sudah ditangani oleh rekan kerja. Muat ulang untuk
              melihat keadaan terbaru.
            instance: /v1/orders/ord_01HZX2Y5K7/fulfilment
    BadRequest:
      description: Malformed or invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/bad-request
            title: Bad request
            status: 400
            detail: Request is malformed or contains invalid syntax.
            instance: /v1/orders
    Unauthorized:
      description: >-
        The token is absent, unverifiable, or expired. The service never trusts
        claims before the signature, issuer, audience, and expiry are verified.
      headers:
        WWW-Authenticate:
          description: OAuth 2.0 / Bearer authentication scheme and error details
          schema:
            type: string
            example: 'Bearer realm="Laundry API", error="invalid_token"'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/unauthorized
            title: Unauthorized
            status: 401
            detail: Authentication is required. Token is missing, invalid, or expired.
            instance: /v1/orders
    Forbidden:
      description: >-
        The token is valid but does not carry the scope this operation requires.
        The scope check runs before any object is loaded, so the refusal leaks
        no information about whether an identifier exists.
      headers:
        WWW-Authenticate:
          description: OAuth 2.0 / Bearer scope error details
          schema:
            type: string
            example: 'Bearer error="insufficient_scope", scope="orders:fulfil"'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/forbidden
            title: Forbidden
            status: 403
            detail: Caller lacks required scope or permission for this operation.
            instance: /v1/orders/ord_01HZX2Y5K7/cancellation
    NotFound:
      description: Resource does not exist or caller lacks authority to view it (identical response to prevent resource existence probing)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/not-found
            title: Resource not found
            status: 404
            detail: Requested resource does not exist or is not accessible.
            instance: /v1/orders/ord_unknown
    Conflict:
      description: Request conflicts with current domain state or idempotency record
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            invalidState:
              value:
                type: https://api.example.com/problems/order-not-cancellable
                title: Order cannot be cancelled
                status: 409
                detail: Order has already entered processing.
                instance: /v1/orders/ord_01HZX2Y5K7/cancellation
            idempotencyConflict:
              value:
                type: https://api.example.com/problems/idempotency-conflict
                title: Idempotency key conflict
                status: 409
                detail: Idempotency-Key was reused with different request data.
                instance: /v1/orders
      headers:
        Retry-After:
          description: Seconds to wait when the request for this key is still processing.
          schema:
            type: integer
            minimum: 1
            example: 5
    UnprocessableContent:
      description: Request data or domain validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/validation-failed
            title: Validation failed
            status: 422
            detail: One or more request fields are invalid.
            instance: /v1/orders
    InternalServerError:
      description: Unexpected internal error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/internal-error
            title: Internal server error
            status: 500
            detail: An unexpected error occurred.
            instance: /v1/orders
    BadGateway:
      description: Upstream dependency failure
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/bad-gateway
            title: Bad gateway
            status: 502
            detail: An upstream dependency returned an invalid response.
            instance: /v1/orders
    ServiceUnavailable:
      description: Service temporarily unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/service-unavailable
            title: Service unavailable
            status: 503
            detail: Service is temporarily unavailable.
            instance: /v1/orders
    GatewayTimeout:
      description: Gateway or upstream timeout
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.example.com/problems/gateway-timeout
            title: Gateway timeout
            status: 504
            detail: Upstream service did not respond in time.
            instance: /v1/orders